+++ LSDLayeredSelectionDisplay/Settings/LSDLayeredSelectionDisplayModSettings.cs
@@ -4,19 +4,25 @@
 
 namespace LSD_Layered_Selection_Display.Settings
 {
+    using System.Collections.Generic;
+    using Colossal;
     using Colossal.IO.AssetDatabase;
     using Game;
     using Game.Input;
     using Game.Modding;
     using Game.Settings;
     using Game.Tools;
+    using Game.UI;
+    using Game.UI.Widgets;
     using LSD_Layered_Selection_Display.Systems;
+    using Newtonsoft.Json.Linq;
     using Unity.Entities;
+    using Unity.Mathematics;
 
     /// <summary>
     /// The mod settings for the LSD Layered Selection Display mod.
     /// </summary>
-    [FileLocation("Mods_LSD_Layered_Selection_Display")]
+    [FileLocation("ModsSettings/LSD_Layered_Selection_Display/LSD_Layered_Selection_Display")]
     public class LSDLayeredSelectionDisplayModSettings : ModSetting
     {
         /// <summary>
@@ -26,7 +32,7 @@ namespace LSD_Layered_Selection_Display.Settings
         public LSDLayeredSelectionDisplayModSettings(IMod mod)
             : base(mod)
         {
-            SetDefaults();
+            // SetDefaults();
         }
 
         /// <summary>
@@ -34,9 +40,22 @@ namespace LSD_Layered_Selection_Display.Settings
         /// </summary>
         public string Version => LSDLayeredSelectionDisplayMod.Instance.Version;
 
+        /// <summary>
+        /// Gets or sets the position of the game list panel.
+        /// </summary>
+        [SettingsUIHidden]
+        public float2 GameListPanelPosition { get; set; }
+
+        /// <summary>
+        /// Gets or sets the position of the editor list panel.
+        /// </summary>
+        [SettingsUIHidden]
+        public float2 EditorListPanelPosition { get; set; }
+
         /// <inheritdoc/>
         public override void SetDefaults()
         {
+            throw new System.NotImplementedException();
         }
     }
 }